Add files via upload
This is going to be my first machine learning project. I will add more data to be scraped after I clean and process what is being collected to this point. I will also break the code up into classes to make it easier to collect data when I scrape it.